refactor: client/components -> client/src/components 로 폴더구조 개선
