+++ apps/client/app/login/page.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { LoginPage } from '../../components/login/page';
+import { LoginPage } from '../../src/components/login/page';
 
 const Page = () => {
   return <LoginPage />;
