+++ apps/client/src/components/section/BodySection.tsx
